Move SeekableData to mca.seekable package
