+++ src/main/java/net/querz/mca/seekable/SeekableByteArray.java
@@ -1,6 +1,5 @@
 package net.querz.mca.seekable;
 
-import net.querz.mca.SeekableData;
 import java.io.*;
 import java.nio.charset.StandardCharsets;
 import java.util.regex.Pattern;
